⚡:Контекст кода для анализа:
Apply AI assistant changes to 4 files.
Ref: ⚡:Контекст кода для анализа:...
